refactor(frontend): move desktop-shell bridge components out of the shell- prefix
components/shell/ holds the layout Shell; shell-accent-reporter,
shell-badge-reporter, and shell-titlebar-strip are Electron desktop-shell
bridge components (importing lib/shell.ts) — an unrelated concept sharing
the prefix. They now live in components/desktop-shell/, matching the
project's own vocabulary. lib/shell.ts and lib/shell-strip.ts are
unchanged. Identifiers untouched — file moves only.
