+++ app/frontend/src/app.tsx
@@ -80,9 +80,9 @@ import { buildServerKillActions } from "@/lib/palette-server-kill";
 import { buildServerProtectActions } from "@/lib/palette-server-protect";
 import { buildShellServerActions } from "@/lib/palette-shell";
 import { canCloseShellWindow, canNewShellWindow, closeShellWindow, isShell, newShellWindow, switchShellServer } from "@/lib/shell";
-import { ShellTitlebarStrip } from "@/components/shell-titlebar-strip";
-import { ShellAccentReporter } from "@/components/shell-accent-reporter";
-import { ShellBadgeReporter } from "@/components/shell-badge-reporter";
+import { ShellTitlebarStrip } from "@/components/desktop-shell/titlebar-strip";
+import { ShellAccentReporter } from "@/components/desktop-shell/accent-reporter";
+import { ShellBadgeReporter } from "@/components/desktop-shell/badge-reporter";
 import { useShellServers } from "@/hooks/use-shell-servers";
 import { readLastPinnedBoard } from "@/lib/last-pinned-board";
 import { buildOpenActions, buildOpenLastUsedAction, buildOpenPrAction } from "@/lib/palette-open";
