+++ app/frontend/src/components/desktop-shell/accent-reporter.tsx
