+++ app/frontend/src/components/desktop-shell/badge-reporter.tsx
