+++ app/frontend/src/components/desktop-shell/titlebar-strip.tsx
