+++ docs/memory/run-kit/desktop-shell.md
@@ -145,7 +145,7 @@ The shell runs **one process with many `BrowserWindow`s**. `app.requestSingleIns
 
 The page's top rung: registered hosts rendered above This Mac, populated by a **one-shot `servers:list` fetch at wire-up** (no poll or subscription — the page is short-lived and list mutations from this page navigate away; the SPA dropdown's open-time-snapshot precedent). The page is already a privileged `servers:*` sender (`isHostsSender` covers the welcome `file://` URL), so list + switch ride the existing `servers` bridge group with **no new IPC channels or preload surfaces** (§ Design Decisions → The welcome hosts rung rides the existing `servers` bridge group). An empty list, an absent `servers` group (older preload), or a failed/malformed answer leaves the section hidden — the first-launch flow is unchanged and no error renders.
 
-Row anatomy mirrors the SPA strip dropdown (`shell-titlebar-strip.tsx`): a 3px left-edge accent bar in the host's persisted `accentColor` (rendered only when the value passes the strict `#RGB`/`#RRGGBB`/`#RRGGBBAA` hex gate — `HOST_ACCENT_HEX` mirrored by value, since the page is a global script that cannot import SPA modules), a fixed-width `✓` marker column for the active host, the name, the dimmed origin (`new URL(url).origin`, falling back to the raw string), and a right-aligned accelerator hint (`⌥⌘{n}` on darwin, `Alt+{n}` elsewhere, list order, capped at 9 — mirroring `hostAcceleratorHint` and the native Hosts menu's binding order). Click or Enter on a row invokes `servers:switch` with that host's id; main's `switchToHost` attaches the view and navigates the window away, so a resolved switch needs no page-side cleanup.
+Row anatomy mirrors the SPA strip dropdown (`desktop-shell/titlebar-strip.tsx`): a 3px left-edge accent bar in the host's persisted `accentColor` (rendered only when the value passes the strict `#RGB`/`#RRGGBB`/`#RRGGBBAA` hex gate — `HOST_ACCENT_HEX` mirrored by value, since the page is a global script that cannot import SPA modules), a fixed-width `✓` marker column for the active host, the name, the dimmed origin (`new URL(url).origin`, falling back to the raw string), and a right-aligned accelerator hint (`⌥⌘{n}` on darwin, `Alt+{n}` elsewhere, list order, capped at 9 — mirroring `hostAcceleratorHint` and the native Hosts menu's binding order). Click or Enter on a row invokes `servers:switch` with that host's id; main's `switchToHost` attaches the view and navigates the window away, so a resolved switch needs no page-side cleanup.
 
 **⇧⌘H is handled locally.** A document `keydown` in `welcome.ts` matches the host-menu chord — `KeyH` + Shift + the platform primary modifier (meta on darwin, ctrl otherwise), exact modifiers, mirroring the SPA registry's `host-menu-open` binding (`lib/keybindings.ts`). With hosts listed, the chord focuses the list's roving seat (roving tabindex: ↓/↑ move without wrapping, Enter selects the focused row, and unmodified `Digit1`–`Digit9` select the Nth row — the listener lives on the list container, so digits never fire while typing in the form inputs); with no hosts listed it focuses the add form's URL field. The handler is shell-owned: no IPC, no accelerator, no coordination with the SPA's binding — the two surfaces never coexist on screen.
 
@@ -458,7 +458,7 @@ The **window-chrome surfaces are hardware-only on every platform** for the same 
 
 The win/linux *menu* contract is nonetheless CI-provable without hardware: the compiled `dist/menu.js` can be loaded under a mocked `electron` module with `process.platform` forced, and the built template's accelerators asserted — that is how the "nothing in the unshifted Ctrl tier" invariant was verified.
 
-The vitest column of the split also covers `palette-shell.test.ts` (the palette-side suite of the `servers` bridge group), plus the SPA halves of the shell chrome — `shell-strip.test.ts`, `shell-titlebar-strip.test.tsx`, and `shell-badge-reporter.test.tsx`.
+The vitest column of the split also covers `palette-shell.test.ts` (the palette-side suite of the `servers` bridge group), plus the SPA halves of the shell chrome — `shell-strip.test.ts`, `desktop-shell/titlebar-strip.test.tsx`, and `desktop-shell/badge-reporter.test.tsx`.
 
 ## Design Decisions
