+++ docs/memory/run-kit/ui/keyboard-and-palette.md
@@ -171,7 +171,7 @@ Macros are also palette-reachable without their key: one `Macro: {label}` entry 
 
 - **`useKeybindingDispatch(handlers)`** — one window-level bubble-phase `keydown` per route shell, keyed by `actionId`. Rules in order: skip `defaultPrevented` events (the shortcuts panel's capture-phase rebind listener claims its own); walk the **enabled** matches in `findMatches` order (§ Scope precedence) and fire the **first one that has a handler at this mount**; when **no** match has a handler the chord falls through untouched (no `preventDefault`), so a route without context for an action leaves browser/pane behavior alone; then `shouldSuppressChord` gates real text inputs unless the binding sets `ignoreInputs`. Bindings and handlers live in refs so the listener registers once per mount and SSE ticks never churn it.
 - **`shouldSuppressChord(target)`** — the single shared input-gating predicate (`lib/keybindings.ts`): suppress only when a real text input owns focus, with two carve-outs — anything inside `.xterm` (the hidden helper textarea is the terminal's normal focus state) and `.rk-chat-input` (the chat lens's analog). jsdom-safe (`contentEditable` attribute fallback for the missing `isContentEditable` getter). Bindings carrying `ignoreInputs` skip it entirely — ⌘K, the shortcuts toggle, `settings-open`, the find/address openers, and the whole surface-toggle family (`compose-toggle`, `zen-toggle`, `sidebar-toggle`, the three tile digits).
-- **Four chords keep their own component-local listeners** rather than joining the dispatcher, because their enablement is local state: ⌘K in `command-palette.tsx` (palette open-state), the `sidebar-toggle` chord (⌘B/⇧Ctrl+B) in `shell.tsx`'s `useSidebarKeyboardToggle`, ⌘. in `app.tsx`'s view-cycle effect (lens availability), and the `host-menu-open` chord (⇧⌘H/⇧Ctrl+H) in `shell-titlebar-strip.tsx` (switcher interactivity + menu open state; the strip's shell-only mount doubles as the handler-presence gate). Each reads its binding through `byAction` and matches with `matchesCombo` + `shouldSuppressChord` (honoring the binding's `ignoreInputs`, mirroring dispatcher rule 4 — `sidebar-toggle` relies on this to fire from the compose textarea) — definition and suppression are the registry's, enablement is the seam's.
+- **Four chords keep their own component-local listeners** rather than joining the dispatcher, because their enablement is local state: ⌘K in `command-palette.tsx` (palette open-state), the `sidebar-toggle` chord (⌘B/⇧Ctrl+B) in `shell.tsx`'s `useSidebarKeyboardToggle`, ⌘. in `app.tsx`'s view-cycle effect (lens availability), and the `host-menu-open` chord (⇧⌘H/⇧Ctrl+H) in `desktop-shell/titlebar-strip.tsx` (switcher interactivity + menu open state; the strip's shell-only mount doubles as the handler-presence gate). Each reads its binding through `byAction` and matches with `matchesCombo` + `shouldSuppressChord` (honoring the binding's `ignoreInputs`, mirroring dispatcher rule 4 — `sidebar-toggle` relies on this to fire from the compose textarea) — definition and suppression are the registry's, enablement is the seam's.
 - **Terminal seam** — `terminal-client.tsx`'s `attachCustomKeyEventHandler` keeps the ⌘C copy case first, then delegates the refusal decision to the pure predicate `shouldRefuseTerminalChord(e, bindings, platform)`, reading both the bindings and the host platform through render-updated refs (the handler attaches once at terminal init, but overrides can change any time). Returning `false` means xterm neither handles nor transmits the key, and xterm does not `preventDefault` a refused key, so the event bubbles to the window dispatcher and the action fires while the terminal owns focus. Three refusal rules:
   - **Any enabled shifted-tier match, on every platform** — legacy TTY encoding cannot distinguish Ctrl+Shift+letter from Ctrl+letter, so xterm would emit the Ctrl-char; refusing costs the pane nothing.
   - **On macOS only, an enabled `cmd`-tier match pressed with `metaKey`** — ⌘ chords never reach the pane as control bytes, so refusal is loss-free and the demoted ⌘[/⌘]/⌘/, ⌘B, ⌘1/⌘2/⌘3, ⌘I, and ⌘↑/⌘↓ (plus shell-host ⌘N/T/W) fire under terminal focus. The `metaKey` gate is load-bearing: `matchesCombo`'s cmd tier also accepts plain Ctrl, and mac Ctrl+[ is ESC.
