+++ docs/memory/run-kit/ui/top-bar.md
@@ -268,7 +268,7 @@ Breadcrumb crumbs advertise their interactivity WITHOUT hover — an affordance 
 
 *Why not hover-only cues*: a `hover:text-text-primary` cue is invisible to first-time users and absent entirely on touch. (`260702-6m46`)
 
-### Desktop-Shell Titlebar Strip (`components/shell-titlebar-strip.tsx` + `lib/shell-strip.ts`)
+### Desktop-Shell Titlebar Strip (`components/desktop-shell/titlebar-strip.tsx` + `lib/shell-strip.ts`)
 
 The instance accent's **third consumer surface**, and the only one that exists outside a browser: inside the Electron viewer shell the native titlebar is hidden, so the SPA draws the window's titlebar itself as a 28px full-width band above the top bar. The accent chain is therefore PWA titlebar meta → top-bar wash/stripe → **shell titlebar strip**, all three fed by the same `useInstanceAccent()` instance.
 
@@ -292,9 +292,9 @@ The instance accent's **third consumer surface**, and the only one that exists o
 
 **Selection hands off to the shell, no optimistic UI.** Picking a row closes the menu and calls `switchShellServer(id)`; the shell's single `switchToHost` seam swaps the displayed host, so there is nothing to update optimistically and no SPA-side path handling exists. The swap is a detach/attach flip between per-host renderer views ([desktop-shell](/run-kit/desktop-shell.md) § Host Views), so this component instance simply stops being displayed — it keeps running, with its state intact, and is shown again unchanged when the user switches back. A `false` resolution surfaces an error toast (the palette's precedent), and selecting the already-active row is a harmless no-op — the same behavior the palette action and the native menu radios have, so all three switch paths behave identically.
 
-**Marker class + mount gating.** While mounted the component sets `rk-shell-strip` on `<html>` (a `useLayoutEffect`, removed on unmount) — the shell's version-skew fallback CSS is keyed on `html:not(.rk-shell-strip)` and self-disables against it ([desktop-shell](/run-kit/desktop-shell.md) § Hidden Titlebar & Accent Strip). `AppLayout` mounts the strip gated on `isShell()`, as a **sibling above** the top-bar wrapper — the top bar itself is untouched, which is the explicit contrast with a titlebar merge. `isShell()` is false in every browser and in Playwright, so **no e2e sees the strip or its dropdown**: the whole surface is covered by three colocated vitest suites — `lib/shell-strip.test.ts` for the pure helpers, `host-form-dialog.test.tsx` for the shared dialog's two modes, and `shell-titlebar-strip.test.tsx` for the component (rendered inside `ToastProvider`, asserting the gate branch, the open-time refetch call count, row anatomy including the accent bar / waiting chip / drag grip, ⌥↑/⌥↓ moves with live hint re-numbering and edge no-ops, drop-commits-one-invoke, the older-shell degradation, shrunk/emptied-list focus behavior, out-of-order resolutions, selection, and Escape/Arrow keys — plus the Disconnect/rename inventory: both icons gated independently on their capabilities, plain rows with unswallowed keys on an older shell, Disconnect-icon click and Delete/Backspace invoking `remove` with the row id (footer unbound), F2/icon opening the prefilled Edit dialog, Save committing only the changed halves (trimmed rename with optimistic update; setUrl normalized to the origin), an invalid URL holding the dialog open on an inline error, unchanged/emptied saves performing no invoke, Save returning focus to the row, the malformed-stored-URL name-only save, the setUrl-less disabled URL field, both failure toasts with refetch, menu-close canceling an open dialog, the `dialogOpen` key-suspension guard, and the remove-shrunk/remove-emptied refetch guards; plus the add-path fork — footer omitted with neither invoker, the `servers:add` page swap on add-only shells (denial toasts), the in-place add-mode dialog on `addDirect` shells (submit-through-addDirect success, inline main-side failure, both-paths preference, menu-close cleanup); plus the ⇧⌘H/digit family — chord open with active-row focus, toggle-close with trigger refocus, non-interactive fall-through, the `HOST_MENU_OPEN_EVENT` seam and its non-interactive inertness, plain-digit selection in accelerator order with beyond-count/modified-digit/dialog-suppression no-ops, and the `:focus-visible` reveal class contract) — plus manual verification in the real shell. The band persists in macOS fullscreen (traffic lights auto-hide over it; no fullscreen bridge flag exists).
+**Marker class + mount gating.** While mounted the component sets `rk-shell-strip` on `<html>` (a `useLayoutEffect`, removed on unmount) — the shell's version-skew fallback CSS is keyed on `html:not(.rk-shell-strip)` and self-disables against it ([desktop-shell](/run-kit/desktop-shell.md) § Hidden Titlebar & Accent Strip). `AppLayout` mounts the strip gated on `isShell()`, as a **sibling above** the top-bar wrapper — the top bar itself is untouched, which is the explicit contrast with a titlebar merge. `isShell()` is false in every browser and in Playwright, so **no e2e sees the strip or its dropdown**: the whole surface is covered by three colocated vitest suites — `lib/shell-strip.test.ts` for the pure helpers, `host-form-dialog.test.tsx` for the shared dialog's two modes, and `desktop-shell/titlebar-strip.test.tsx` for the component (rendered inside `ToastProvider`, asserting the gate branch, the open-time refetch call count, row anatomy including the accent bar / waiting chip / drag grip, ⌥↑/⌥↓ moves with live hint re-numbering and edge no-ops, drop-commits-one-invoke, the older-shell degradation, shrunk/emptied-list focus behavior, out-of-order resolutions, selection, and Escape/Arrow keys — plus the Disconnect/rename inventory: both icons gated independently on their capabilities, plain rows with unswallowed keys on an older shell, Disconnect-icon click and Delete/Backspace invoking `remove` with the row id (footer unbound), F2/icon opening the prefilled Edit dialog, Save committing only the changed halves (trimmed rename with optimistic update; setUrl normalized to the origin), an invalid URL holding the dialog open on an inline error, unchanged/emptied saves performing no invoke, Save returning focus to the row, the malformed-stored-URL name-only save, the setUrl-less disabled URL field, both failure toasts with refetch, menu-close canceling an open dialog, the `dialogOpen` key-suspension guard, and the remove-shrunk/remove-emptied refetch guards; plus the add-path fork — footer omitted with neither invoker, the `servers:add` page swap on add-only shells (denial toasts), the in-place add-mode dialog on `addDirect` shells (submit-through-addDirect success, inline main-side failure, both-paths preference, menu-close cleanup); plus the ⇧⌘H/digit family — chord open with active-row focus, toggle-close with trigger refocus, non-interactive fall-through, the `HOST_MENU_OPEN_EVENT` seam and its non-interactive inertness, plain-digit selection in accelerator order with beyond-count/modified-digit/dialog-suppression no-ops, and the `:focus-visible` reveal class contract) — plus manual verification in the real shell. The band persists in macOS fullscreen (traffic lights auto-hide over it; no fullscreen bridge flag exists).
 
-### Desktop-Shell Waiting Badge Reporter (`components/shell-badge-reporter.tsx`)
+### Desktop-Shell Waiting Badge Reporter (`components/desktop-shell/badge-reporter.tsx`)
 
 A null-rendering subscriber, mounted in `AppLayout` beside the strip and gated on `isShell()`, that mirrors the waiting-agent count to the OS dock/taskbar badge — the attention tier escaping the page (§ Attention Surfacing). It is the fifth `lib/waiting.ts` consumer and the only one whose output leaves the browser.
