+++ docs/memory/run-kit/ui/visual-design.md
@@ -227,7 +227,7 @@ A per-instance accent color makes multiple run-kit instances (laptop, Mac mini, 
 
 **Provider + consumers** — `InstanceAccentProvider` / `useInstanceAccent()` (`contexts/instance-accent-context.tsx`) mounts once in `RootWrapper` inside `ThemeProvider` > `ToastProvider` and above `ChromeProvider`, so it can read `useTheme()` for the theme-derived hexes and `useToast()` for the write-failure toast. It fetches `getInstanceColor()` + `getHealth()` once on mount (StrictMode-guarded via a `didFetchRef`), seeds the first paint from the echo, resolves per the chain, derives `{stripeHex, washHex, titlebarHex}` from the active theme, keeps the echo + theme-color meta in sync on every accent/theme change, and exposes all three on the context value (`{color, isExplicit, stripeHex, washHex, titlebarHex, setColor}`). `titlebarHex` is nullable like its siblings and has two consumers: the theme-color meta bridge effect, and the desktop shell's titlebar strip, which renders from it directly rather than re-deriving the same blend (§ Desktop-Shell Titlebar Strip). `setColor(descriptor | null)` updates state optimistically then POSTs via `setInstanceColor`, toasting on failure. Every rendering surface reads from this ONE instance — one fetch, one state, so a pick in the HOST panel repaints the top-bar stripe without a reload. A test seam `InstanceAccentValueProvider` injects a fixed value (mirroring the session-context value-injection pattern).
 
-**Theme-derived hexes** (`deriveAccentHexes(value, theme)` in `src/instance-accent.ts`, never hardcoded) — a three-hex split, each a distinct accent surface: `stripeHex` is the **contrast-guarded** family hex — `colorValueToHex(value, palette)` then `adjustBorderForContrast(src, bg, isDark, BORDER_MIN_CONTRAST)` (the same guarded-hex derivation server tiles/stripes use, § Color Tinting) — used for the 2px top-bar stripe and the HOST hostname tint, and, in-shell, reported over `accent:set` by the render-nothing `ShellAccentReporter` (`components/shell-accent-reporter.tsx`, mounted beside the badge reporter behind `isShell()`, ref-guarded change-only reporting, null never reported) so the shell persists the full-strength value for the host-switcher's edge bars rather than the meta's 35% blend. `washHex` is a subtle `blendHex(src, bg, INSTANCE_WASH_RATIO = 0.065)` (~6.5%) accent-into-background blend for the top-bar background wash. `titlebarHex` is the strongest of the three — `blendHex(src, bg, INSTANCE_TITLEBAR_RATIO = 0.35)` (~35%) accent-into-background — and paints **every titlebar-class surface**: the theme-color meta content (the installed-PWA titlebar tint) and the desktop shell's page-drawn titlebar strip, in both cases a tinted band with the full-brightness `stripeHex` 2px stripe visible directly below it. `INSTANCE_TITLEBAR_RATIO` sits beside `INSTANCE_WASH_RATIO` (a taste constant). Blend descriptors (`"1+3"`) render through the same `resolveFamily`-based owned-family mapping as single indices — no new blend scheme. All three recompute on a theme switch (light themes get a light-background blend by construction); nothing hardcoded survives. An unrecognized descriptor yields `null` (no partial object).
+**Theme-derived hexes** (`deriveAccentHexes(value, theme)` in `src/instance-accent.ts`, never hardcoded) — a three-hex split, each a distinct accent surface: `stripeHex` is the **contrast-guarded** family hex — `colorValueToHex(value, palette)` then `adjustBorderForContrast(src, bg, isDark, BORDER_MIN_CONTRAST)` (the same guarded-hex derivation server tiles/stripes use, § Color Tinting) — used for the 2px top-bar stripe and the HOST hostname tint, and, in-shell, reported over `accent:set` by the render-nothing `ShellAccentReporter` (`components/desktop-shell/accent-reporter.tsx`, mounted beside the badge reporter behind `isShell()`, ref-guarded change-only reporting, null never reported) so the shell persists the full-strength value for the host-switcher's edge bars rather than the meta's 35% blend. `washHex` is a subtle `blendHex(src, bg, INSTANCE_WASH_RATIO = 0.065)` (~6.5%) accent-into-background blend for the top-bar background wash. `titlebarHex` is the strongest of the three — `blendHex(src, bg, INSTANCE_TITLEBAR_RATIO = 0.35)` (~35%) accent-into-background — and paints **every titlebar-class surface**: the theme-color meta content (the installed-PWA titlebar tint) and the desktop shell's page-drawn titlebar strip, in both cases a tinted band with the full-brightness `stripeHex` 2px stripe visible directly below it. `INSTANCE_TITLEBAR_RATIO` sits beside `INSTANCE_WASH_RATIO` (a taste constant). Blend descriptors (`"1+3"`) render through the same `resolveFamily`-based owned-family mapping as single indices — no new blend scheme. All three recompute on a theme switch (light themes get a light-background blend by construction); nothing hardcoded survives. An unrecognized descriptor yields `null` (no partial object).
 
 **Top-bar stripe + wash** (`AppLayout`, `app.tsx`) — the persistent top bar's `shrink-0` wrapper (above `RootTopBar`) carries a 2px accent stripe (an `aria-hidden` `<div style={{height:"2px", backgroundColor: stripeHex}}>`) plus the `washHex` as its `backgroundColor` (the `TopBar` header has no background of its own, so the wash on the wrapper shows through). Both render only when their hex is non-null (no accent resolved ⇒ neither renders). This is the "which instance" channel living where the sidebar colors do not.
